refector base64/hex checks within lightfuzz + tests
